Task/new context algorithm 01 (#264)
* Step 01 ready - contexts are created as hashtables or arrays of contexts, for some cases

* preparing to merge with develop

* Initial download and creation of Core Context any any other preloaded context

* Implemented the new context algorithm

* Fixed a bug in dot-for-eq-sign that destroyed the context-cache (attr-names inside the cache has '=' instead of '.')

* Last error fixed (new notifications didn't work due to ./=)

* style

* Renamed OrionldAltContext -> OrionldContext

* Copyright headers

* Refactor - Core context (and default URL) to its own module. OrionldContext to its own module

* More refactoring - orionldContextCache in its own module

* More refactoring: new modules: orionldContextDownload, orionldContextHashTablesFill, orionldContextPresent, orionldContextTreePresent

* Finished the refractor - split orionld/context into a total of 26 modules

* removed safe-guarded old implementation of the @context algorithm

* Added khash to docker files

* Removed an empty line in a functest

+++ src/lib/orionld/common/qLex.cpp
@@ -52,7 +52,7 @@ static QNode* qStringPush(QNode* prev, char* stringValue)
 //
 // qDateTimePush -
 //
-static QNode* qDateTimePush(QNode* prev, long long dateTime)
+static QNode* qDateTimePush(QNode* prev, int64_t dateTime)
 {
   QNode* qNodeP = qNode(QNodeIntegerValue);
 
@@ -167,7 +167,7 @@ static QNode* qTermPush(QNode* prev, char* term, char** titleP, char** detailsP)
 
     if (type == QNodeIntegerValue)
     {
-      long long dTime;
+      int64_t dTime;
 
       if (dateTime == true)
       {
@@ -400,8 +400,8 @@ QNode* qLex(char* s, char** titleP, char** detailsP)
 
       *sP = 0;
       ++sP;
-      long long           dateTime;
-      unsigned long long  sLen = (unsigned long long) (sP - start - 2);
+      int64_t   dateTime;
+      uint64_t  sLen = (uint64_t) (sP - start - 2);
 
       if ((sLen > 4) && (start[4] == '-') && ((dateTime = parse8601Time(start)) != -1))
         current = qDateTimePush(current, dateTime);
